+++ src/test/scala/edu/rit/cs/mmior/jsonoid/discovery/DefinitionTransformerSpec.scala
@@ -41,7 +41,6 @@ class DefinitionTransformerSpec extends UnitSpec {
     (props \ "quux").extract[Map[String, String]] shouldBe ref
 
     // Check that a reference object was defined
-    @SuppressWarnings(Array("org.wartremover.warts.OptionPartial"))
     val referencedSchema = transformedSchema
       .findByPointer("/foo")
       .get